♻️ Refactor: dummy-files 분리
